+++ src/commands/buildUpload.ts
@@ -1,10 +1,5 @@
 import * as vscode from "vscode";
-import * as child_process from "child_process";
-import { promisify } from "util";
-
-import { parseMakeOutput } from "./cli-parsing";
-import { output } from "../extension";
-import { getChildProcessPath } from "../one-click/path";
+import { BaseCommand } from "./base-command";
 /**
  * Call the PROS build CLI command.
  *
@@ -12,45 +7,13 @@ import { getChildProcessPath } from "../one-click/path";
  */
 
 const runBuildUpload = async () => {
-  await vscode.window.withProgress(
-    {
-      location: vscode.ProgressLocation.Notification,
-      title: "Building and Uploading Project",
-      cancellable: false,
-    },
-    async (progress, token) => {
-      try {
-        // Command to run to build and upload project
-        var command = `pros mu --project "${vscode.workspace.workspaceFolders?.[0].uri.fsPath}" --machine-output ${process.env["PROS_VSCODE_FLAGS"]}`;
-        console.log(command);
-        const { stdout, stderr } = await promisify(child_process.exec)(
-          command,
-          {
-            env: {
-              ...process.env,
-              // eslint-disable-next-line @typescript-eslint/naming-convention
-              PATH: getChildProcessPath(),
-            },
-            maxBuffer: 1024 * 1024 * 50,
-          }
-        );
-        await vscode.window.showInformationMessage("Project Built!");
-      } catch (error: any) {
-        if (!error.stdout.includes("No v5 ports")) {
-          const rtn = await vscode.window.showErrorMessage(
-            parseMakeOutput(error.stdout),
-            "View Output!",
-            "No Thanks!"
-          );
-          if (rtn === "View Output!") {
-            output.show();
-          }
-        } else {
-          vscode.window.showErrorMessage(parseMakeOutput(error.stdout));
-        }
-      }
-    }
-  );
+  const buildUploadCommand = new BaseCommand({
+    command: "pros",
+    args: ["mu", ...`${process.env.PROS_VSCODE_FLAGS}`.split(" ")],
+    message: "Building and Uploading Project",
+    requiresProsProject: true,
+  });
+  await buildUploadCommand.runCommand();
 };
 
 export const buildUpload = async () => {