+++ src/commands/index.ts
@@ -10,3 +10,7 @@ export * from "./stop";
 export * from "./runmedic";
 export * from "./fw-update";
 export * from "./prosDocs";
+export * from "./command_tools";
+export * from "./base-command";
+export * from "./set-team-number";
+export * from "./set-robot-name";