+++ src/one-click/install.ts
@@ -84,8 +84,8 @@ export async function uninstall(context: vscode.ExtensionContext) {
       },
       async (progress, token) => {
         let promises: Promise<any>[] = [];
-        promises.push(removeDirAsync(path.join(globalPath, "install"), true));
-        promises.push(removeDirAsync(path.join(globalPath, "download"), true));
+        promises.push(removeDirAsync(path.join(globalPath, "install"), false));
+        promises.push(removeDirAsync(path.join(globalPath, "download"), false));
         await Promise.all(promises);
       }
     );
@@ -93,10 +93,15 @@ export async function uninstall(context: vscode.ExtensionContext) {
   }
 }
 
-async function getUrls(cliVersion: number, toolchainVersion: string) {
+async function getUrls(
+  cliVersion: number,
+  toolchainVersion: string,
+  vexcomVersion: string
+) {
   var downloadCli = `https://github.com/purduesigbots/pros-cli/releases/download/${cliVersion}/pros_cli-${cliVersion}-lin-64bit.zip`;
   var downloadToolchain =
     "https://developer.arm.com/-/media/Files/downloads/gnu-rm/10.3-2021.10/gcc-arm-none-eabi-10.3-2021.10-x86_64-linux.tar.bz2";
+  var downloadVexcom = `https://pros.cs.purdue.edu/v5/_static/releases/vexcom_${vexcomVersion}-linux-x64.zip`;
 
   await prosLogger.log(
     "OneClick",
@@ -107,12 +112,14 @@ async function getUrls(cliVersion: number, toolchainVersion: string) {
     // Set system, path seperator, and downloads to windows version
     downloadCli = `https://github.com/purduesigbots/pros-cli/releases/download/${cliVersion}/pros_cli-${cliVersion}-win-64bit.zip`;
     downloadToolchain = `https://github.com/purduesigbots/toolchain/releases/download/${toolchainVersion}/pros-toolchain-windows.zip`;
+    downloadVexcom = `https://pros.cs.purdue.edu/v5/_static/releases/vexcom_${vexcomVersion}-win32.zip`;
   } else if (getOperatingSystem() === "macos") {
     await prosLogger.log("OneClick", `MacOS detected, using MacOS URLS`);
     // Set system, path seperator, and downloads to windows version
     downloadCli = `https://github.com/purduesigbots/pros-cli/releases/download/${cliVersion}/pros_cli-${cliVersion}-macos-64bit.zip`;
     downloadToolchain =
       "https://developer.arm.com/-/media/Files/downloads/gnu-rm/10.3-2021.10/gcc-arm-none-eabi-10.3-2021.10-mac.tar.bz2";
+    downloadVexcom = `https://pros.cs.purdue.edu/v5/_static/releases/vexcom_${vexcomVersion}-osx.zip`;
     os.cpus().some((cpu) => {
       if (cpu.model.includes("Apple M")) {
         downloadCli = `https://github.com/purduesigbots/pros-cli/releases/download/${cliVersion}/pros_cli-${cliVersion}-macos-arm64bit.zip`;
@@ -168,7 +175,7 @@ async function getUrls(cliVersion: number, toolchainVersion: string) {
     }
   }
 
-  return [downloadCli, downloadToolchain];
+  return [downloadCli, downloadToolchain, downloadVexcom];
 }
 
 export async function install(context: vscode.ExtensionContext) {
@@ -183,22 +190,34 @@ export async function install(context: vscode.ExtensionContext) {
     "OneClick",
     "Configuring Environment Variables for PROS"
   );
-  //await configurePaths(context);
+  await configurePaths(context);
   const globalPath = context.globalStorageUri.fsPath;
   await prosLogger.log("OneClick", "Fetching Operating System....");
   const system = getOperatingSystem();
   await prosLogger.log("OneClick", `Operating System Detected: ${system}`);
 
   await prosLogger.log("OneClick", "Fetching Latest CLI Version....");
-  var cliVersion = await getCurrentReleaseVersion(
-    "https://api.github.com/repos/purduesigbots/pros-cli/releases/latest"
-  );
-  let releaseVersionNumber = +cliVersion.replace(/\./gi, "") ?? 0;
-  const toolchainVersion = await getCurrentReleaseVersion(
-    "https://api.github.com/repos/purduesigbots/toolchain/releases/latest"
-  );
 
-  await prosLogger.log("OneClick", `CLI Version: ${cliVersion}`);
+  let cliVersion, releaseVersionNumber, toolchainVersion;
+  try {
+    cliVersion = await getCurrentReleaseVersion(
+      "https://api.github.com/repos/purduesigbots/pros-cli/releases/latest"
+    );
+    releaseVersionNumber = +cliVersion.replace(/\./gi, "") ?? 0;
+    toolchainVersion = await getCurrentReleaseVersion(
+      "https://api.github.com/repos/purduesigbots/toolchain/releases/latest"
+    );
+    await prosLogger.log("OneClick", `CLI Version: ${cliVersion}`);
+  } catch (e: any) {
+    await prosLogger.log("OneClick", "Failed to access version number");
+    console.log("Hit the rate limit, please try again after some time.");
+    cliVersion = undefined;
+  }
+  let vexcomVersion = "1_0_0_23";
+
+  if (cliVersion === undefined || toolchainVersion === undefined) {
+    throw new Error("Failed to access version number");
+  }
 
   // Get system type, path string separator, CLI download url, and toolchain download url.
   // Default variables are based on linux.
@@ -206,9 +225,10 @@ export async function install(context: vscode.ExtensionContext) {
     "OneClick",
     "Fetching CLI and Toolchain Download URLs...."
   );
-  let [downloadCli, downloadToolchain] = await getUrls(
+  let [downloadCli, downloadToolchain, downloadVexcom] = await getUrls(
     cliVersion,
-    toolchainVersion
+    toolchainVersion,
+    vexcomVersion
   );
   await prosLogger.log("OneClick", `CLI Download URL: ${downloadCli}`);
   await prosLogger.log(
@@ -218,10 +238,14 @@ export async function install(context: vscode.ExtensionContext) {
 
   // Set the installed file names
   var cliName = `pros-cli-${system}.zip`;
+  var vexcomName = `vex-vexcom-${system}.zip`;
   // Title of prompt depending on user's installed CLI
+  let [cliExecPath] = getIntegratedTerminalPaths(context);
+  const addQuotes =
+    getOperatingSystem() === "macos" && !os.cpus()[0].model.includes("Apple M");
   var title = await getInstallPromptTitle(
     path.join(
-      `"${path.join(globalPath, "install", `pros-cli-${system}`)}"`,
+      `${addQuotes ? `"` : ""}${cliExecPath}${addQuotes ? `"` : ""}`,
       "pros"
     ),
     releaseVersionNumber ?? 0
@@ -234,6 +258,10 @@ export async function install(context: vscode.ExtensionContext) {
     })) ?? false;
   const toolchainWorking =
     (await verifyToolchain().catch((err) => {
+      console.log(err);
+    })) ?? false;
+  const vexcomWorking =
+    (await verifyVexcom().catch((err) => {
       console.log(err);
     })) ?? false;
 
@@ -256,9 +284,19 @@ export async function install(context: vscode.ExtensionContext) {
     }`,
     toolchainWorking ? "INFO" : "WARNING"
   );
+  await prosLogger.log(
+    "OneClick",
+    `${
+      vexcomWorking
+        ? "Vexcom appears to be functional"
+        : "Vexcom not functional or not installed"
+    }`,
+    vexcomWorking ? "INFO" : "WARNING"
+  );
 
   console.log("CLI Working: " + cliWorking);
   console.log("Toolchain Working: " + toolchainWorking);
+  console.log("Vexcom Working: " + vexcomWorking);
   // Name of toolchain download depending on system
   var toolchainName = `pros-toolchain-${
     system === "windows" ? `${system}.zip` : `${system}.tar.bz2`
@@ -285,12 +323,14 @@ export async function install(context: vscode.ExtensionContext) {
       " | cliWorking: " +
       cliWorking +
       " | toolchainWorking: " +
-      toolchainWorking
+      toolchainWorking +
+      " | vexcomWorking: " +
+      vexcomWorking
   );
   //if everything works and cli is up to date, do nothing
-  if (cliWorking && toolchainWorking && cliUpToDate) {
+  if (cliWorking && toolchainWorking && cliUpToDate && vexcomWorking) {
     // tell the user that everything is up to date
-    await vscode.window.showInformationMessage(
+    vscode.window.showInformationMessage(
       "CLI and Toolchain currently working and up to date."
     );
     console.log("Everything is up to date");
@@ -303,7 +343,7 @@ export async function install(context: vscode.ExtensionContext) {
   }
 
   // if CLI is up to date but toolchain is not working, install just the toolchain
-  else if (cliUpToDate && cliWorking && !toolchainWorking) {
+  else if (cliUpToDate && cliWorking && vexcomWorking && !toolchainWorking) {
     let prompttitle = "PROS Toolchain is not working. Install now?";
     console.log(prompttitle);
     let labelResponse = await vscode.window.showInformationMessage(
@@ -321,7 +361,7 @@ export async function install(context: vscode.ExtensionContext) {
       );
       console.log("removing directory " + deleteDir);
       await prosLogger.log("OneClick", "removing directory " + deleteDir);
-      await removeDirAsync(deleteDir, true).catch((e) => {
+      await removeDirAsync(deleteDir, false).catch((e) => {
         console.log(e);
       });
       await prosLogger.log(
@@ -337,7 +377,10 @@ export async function install(context: vscode.ExtensionContext) {
       return;
     }
     // if the toolchain is working but the cli is not working or out of date, install just the cli
-  } else if (toolchainWorking && !(cliWorking && cliUpToDate)) {
+  } else if (
+    toolchainWorking &&
+    !(cliWorking && cliUpToDate && vexcomWorking)
+  ) {
     const prompttitle = `PROS CLI is ${
       cliWorking ? "out of date. Update" : "not working. Install"
     } now?`;
@@ -358,14 +401,27 @@ export async function install(context: vscode.ExtensionContext) {
       );
       console.log("removing directory " + deleteDir);
       await prosLogger.log("OneClick", "removing directory " + deleteDir);
-      await removeDirAsync(deleteDir, true).catch((e) => {
+      await removeDirAsync(deleteDir, false).catch((e) => {
+        console.log(e);
+      });
+      await removeDirAsync(
+        path.join(
+          context.globalStorageUri.fsPath,
+          "install",
+          "`vex-vexcom-${system}`"
+        ),
+        false
+      ).catch((e) => {
         console.log(e);
       });
       await prosLogger.log(
         "OneClick",
         "CLI is not working. Installing just the CLI"
       );
-      promises = [downloadextract(context, downloadCli, cliName)];
+      promises = [
+        downloadextract(context, downloadCli, cliName),
+        downloadextract(context, downloadVexcom, vexcomName),
+      ];
     } else {
       await prosLogger.log(
         "OneClick",
@@ -391,7 +447,7 @@ export async function install(context: vscode.ExtensionContext) {
       console.log("removing old cli and toolchain");
       await removeDirAsync(
         path.join(context.globalStorageUri.fsPath, "install"),
-        true
+        false
       ).catch((e) => {
         console.log(e);
       });
@@ -409,6 +465,7 @@ export async function install(context: vscode.ExtensionContext) {
       promises = [
         downloadextract(context, downloadCli, cliName),
         downloadextract(context, downloadToolchain, toolchainName),
+        downloadextract(context, downloadVexcom, vexcomName),
       ];
     } else {
       await prosLogger.log(
@@ -465,7 +522,7 @@ export async function cleanup(
       title: "Verifying Installation",
       cancellable: true,
     },
-    async (progress, token) => {
+    async () => {
       try {
         const globalPath = context.globalStorageUri.fsPath;
         await prosLogger.log(
@@ -477,18 +534,16 @@ export async function cleanup(
             prosLogger.log("OneClick", e, "ERROR");
           }
         );
-        await prosLogger.log("OneClick", `Configuring environment variables`);
-        await configurePaths(context).catch((e) => {
-          prosLogger.log("OneClick", e, "ERROR");
-        });
+
         await prosLogger.log(
           "OneClick",
           `Verifying that CLI and Toolchain are working`
         );
-
         await chmod(globalPath, system);
-
-        //await configurePaths(context);
+        await prosLogger.log("OneClick", `Configuring environment variables`);
+        await configurePaths(context).catch((e) => {
+          prosLogger.log("OneClick", e, "ERROR");
+        });
 
         // Ensure that toolchain and cli are working
         let cliSuccess =
@@ -499,14 +554,18 @@ export async function cleanup(
           (await verifyToolchain().catch((err) => {
             prosLogger.log("OneClick", err, "ERROR");
           })) ?? false;
-        if (cliSuccess && toolchainSuccess) {
+        let vexcomSuccess =
+          (await verifyVexcom().catch((err) => {
+            prosLogger.log("OneClick", err, "ERROR");
+          })) ?? false;
+        if (cliSuccess && toolchainSuccess && vexcomSuccess) {
           vscode.window.showInformationMessage(
             "CLI and Toolchain are working!"
           );
         } else {
           vscode.window.showErrorMessage(
-            `${cliSuccess ? "" : "CLI"}
-            ${!cliSuccess && !toolchainSuccess ? " and " : ""}
+            `${cliSuccess && vexcomSuccess ? "" : "CLI"}
+            ${!cliSuccess && !toolchainSuccess && !vexcomSuccess ? " and " : ""}
             ${toolchainSuccess ? "" : "Toolchain"}
             Installation Failed!`
           );
@@ -527,7 +586,8 @@ export async function configurePaths(
   repeat: boolean = true
 ) {
   await prosLogger.log("OneClick", "Getting paths for integrated terminal");
-  let [cliExecPath, toolchainPath] = getIntegratedTerminalPaths(context);
+  let [cliExecPath, toolchainPath, vexcomPath] =
+    getIntegratedTerminalPaths(context);
 
   // return if the path is already configured
   const addQuotes =
@@ -555,6 +615,7 @@ export async function configurePaths(
   if (PATH_SEP === ":") {
     cliExecPath = cliExecPath.replace(/\\/g, "");
     toolchainPath = toolchainPath.replace(/\\/g, "");
+    vexcomPath = vexcomPath.replace(/\\/g, "");
   }
   TOOLCHAIN = process.env["PROS_TOOLCHAIN"] ?? toolchainPath;
   // Set CLI environmental variable file location
@@ -567,14 +628,20 @@ export async function configurePaths(
     process.env["PATH"]
       ?.split(PATH_SEP)
       .filter((x) => x.includes(toolchainPath)).length ?? 0;
+  let pathVexcomCount: number =
+    process.env["PATH"]?.split(PATH_SEP).filter((x) => x.includes(vexcomPath))
+      .length ?? 0;
 
   prosLogger.log("OneClick", `CLI path count: ${pathCliCount}`);
   prosLogger.log("OneClick", `Toolchain path count: ${pathToolchainCount}`);
+  prosLogger.log("OneClick", `Vexcom path count: ${pathVexcomCount}`);
   console.log(`CLI path count: ${pathCliCount}`);
   console.log(`Toolchain path count: ${pathToolchainCount}`);
+  console.log(`Vexcom path count: ${pathVexcomCount}`);
   if (
     pathCliCount > 2 &&
     pathToolchainCount > 2 &&
+    pathVexcomCount > 2 &&
     process.env["PROS_TOOLCHAIN"]?.includes(TOOLCHAIN)
   ) {
     console.log("path already configured");
@@ -587,12 +654,13 @@ export async function configurePaths(
   await prosLogger.log("OneClick", process.env.PATH ?? "no PATH", "INFO");
   process.env.PATH = `${process.env.PATH}`; // bypass compile errors
   await prosLogger.log("OneClick", process.env.PATH ?? "no PATH", "INFO");
-  process.env.PATH = `${
-    addQuotes ? `"` : ""
-  }${cliExecPath}${PATH_SEP}${path.join(
-    toolchainPath,
-    "bin"
-  )}${PATH_SEP}${process.env.PATH.replace(/\"/g, "")}${addQuotes ? `"` : ""}`;
+  process.env.PATH =
+    `${addQuotes ? `"` : ""}` +
+    `${cliExecPath}${PATH_SEP}` +
+    `${path.join(toolchainPath, "bin")}${PATH_SEP}` +
+    `${vexcomPath}${PATH_SEP}` +
+    `${process.env.PATH.replace(/\"/g, "")}` +
+    `${addQuotes ? `"` : ""}`;
   await prosLogger.log("OneClick", process.env.PATH ?? "no PATH", "INFO");
   // Make PROS_TOOCLHAIN variable
   await prosLogger.log("OneClick", "Setting PROS_TOOLCHAIN");
@@ -664,4 +732,28 @@ async function verifyToolchain() {
     console.log(stderr);
   }
   return stdout.replace(".exe", "").startsWith(`arm-none-eabi-g++ (G`);
+}
+
+async function verifyVexcom() {
+  await prosLogger.log("OneClick", "Verifying VEXCOM");
+  var command = "vexcom --version";
+
+  const { stdout, stderr } = await promisify(child_process.exec)(command, {
+    timeout: 5000,
+    env: {
+      ...process.env,
+      // eslint-disable-next-line @typescript-eslint/naming-convention
+      PATH: getChildProcessPath(),
+    },
+  });
+
+  if (stderr) {
+    await prosLogger.log(
+      "OneClick",
+      `VEXCOM verification failed with error ${stderr}`,
+      "error"
+    );
+    console.log(stderr);
+  }
+  return stdout.replace(".exe", "").startsWith("vexcom: version");
 }